The first connection between books and people
The introduction of `FernandUser`

FernandUser’s unique identifier is an email address.
The code to have an email address based Django User model is taken from
https://github.com/Liberationtech/django-libtech-emailuser

The introduction of `Collaboration`

A book has collaborations, which is a connection between the book,
a FernandUser, and a Role, i.e. Author, Editor, Bookbinder…